feature(unicode): Implement UTF16 support with ICU4C.

+++ Code/Commando/apppacketstats.cpp
@@ -401,7 +401,7 @@ cAppPacketStats::Get_Description
 	char last = ' ';
 	char next = ' ';
 	char * p = WorkingString.Peek_Buffer();
-	for (int i = 0; i < WorkingString.Get_Length(); i++)
+	for (size_t i = 0; i < WorkingString.Get_Length(); i++)
 	{
 		if (i == 0)
 		{
